django-yamlconf: Update VERSION definition/handling
The version info for the package was defined as simple variables
in the __init__.py file.  This introduced dependencies on third
party library during installation (six) that might not be present
(the version was being check before the dependencies are installed).

This change moves the version info to a VERSION file and both the
__init__.py and setup.py files have been updated to used the
triple defined there as the version number (major, minor, patch).

Tests Performed: "make check" OK