django-yamlconf: Restructured documentation a little
Updated the "index.rst" to make it a regular committed file (rather
than a file generated from the README.md file, via "pandoc").  Moved
the documentation on the public methods to the doc comments on the
methods and used "autofunction" to pull it into the "index.rst" file.

Will continue to restructure into separate file and reduce the
README.md file (replace most of the documentation there with links
to readthedocs.org).

Tests Performed: Rebuilt the docs ("make html").