django-yamlconf: Extend "defined_attributes" method
Extended the "defined_attributes" method my merging in the functionality
of "sf_attributes" (sysfiles attributes).  This merging occurs if the
keyword argument "template_use" is True and pulls in "top-level"
dictionary values from the settings file making them available when
the attributes are used in a template file.

This functionality was prompted by the need to serve one of the system
template files over HTTP (allowing clients to update their config files
based on a pull from a central server).

Tests Performed: "make check": 54 passed, 0 failures