Updated requirements to fix security issues
GitHub Dependabot now raises an issues if a package with
security issues, i.e., Django, does not have an explicit
version defined in requirements files for cross checking:
```
Dependabot can't update vulnerable dependencies without a lockfile

The currently installed version can't be determined.

To resolve the issue add a supported lockfile (Pipfile.lock, pyproject.lock or poetry.lock).
```
The requirements files here are for development or as
part of an example application.

To silence Dependabot the package rule for Django was
changed from

	Django<4.0

to

	Django==3.2.18

The latest 3.2 version available.  Will now need to make
updates as security fixes for 3.2 series are released or
simply rename my requirements files.